+++ CHANGES.md
@@ -1,3 +1,7 @@
+## next
+
+ * Add `HashMap.findWithDefault` (deprecates `HashMap.lookupDefault`)
+
 ## 0.2.10.0
 
  * Add `HashMap.alterF`.